pyproject.toml
[tool.poetry]
name = "langchain-sample-app"
version = "0.1.0"
description = ""
authors = ["masatotezuka <tez.0731.mst@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "./" }
]
[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
langchain-openai = "^0.2.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
